requirements.txt
aiohttp>=3.9.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pyyaml>=6.0.0
playwright>=1.40.0
reportlab>=4.0.0
python-docx>=1.1.0
pyshark>=0.6.0
scapy>=2.5.0
boto3>=1.34.0
semgrep>=1.155.0